Crate iroh_mainline_address_lookup

Pkarr based address lookup for iroh, supporting both relay servers and the DHT.

This module contains pkarr-based address lookup for iroh which can use both pkarr relay servers as well as the Mainline DHT directly. See the pkarr module for an overview of pkarr.

Structs

Builder

Builder for DhtAddressLookup.

DhtAddressLookup

Pkarr Mainline DHT and relay server address lookup.